# playerManagementSystem/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a player management system server.",
        "title": "Player Management System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Steven Poon",
            "url": "https://github.com/RYANCOAL9999",
            "email": "lmf242003@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/levels": {
            "get": {
                "description": "Retrieve a list of levels from the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "levels"
                ],
                "summary": "List levels",
                "responses": {
                    "200": {
                        "description": "A list of levels",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Level"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create a new level in the database using the provided level details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "levels"
                ],
                "summary": "Create a new level",
                "parameters": [
                    {
                        "description": "Level details to be created",
                        "name": "level",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Level"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Level created successfully with the generated ID",
                        "schema": {
                            "type": "number"
                        }
                    },
                    "400": {
                        "description": "Bad request due to invalid input",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/players": {
            "get": {
                "description": "Retrieve a list of players and their ranks from the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "List players",
                "responses": {
                    "200": {
                        "description": "A list of players with their ranks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PlayerRank"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update the details of an existing player in the database using the provided player information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Update player details",
                "parameters": [
                    {
                        "description": "Player details to be updated",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PlayerRank"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request due to invalid input",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create a new player in the database using the provided player details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Create a new player",
                "parameters": [
                    {
                        "description": "Player details to be created",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PlayerRank"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Player created successfully with the generated ID",
                        "schema": {
                            "type": "number"
                        }
                    },
                    "400": {
                        "description": "Bad request due to invalid input",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/players/{id}": {
            "get": {
                "description": "Get details of a specific player identified by their ID from the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Retrieve a player by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player details",
                        "schema": {
                            "$ref": "#/definitions/models.PlayerRank"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Remove a player from the database using the provided player ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Delete a player",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Level": {
            "type": "object",
            "required": [
                "lv",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lv": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.PlayerRank": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lv": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}